Remove generic parsers (#24)
